transcriber/requirements.txt
youtube_transcript_api
pathlib
google-generativeai 
streamlit 
python-dotenv